drivers/DRM:Rename the drm driver to be compatible with hikey960 and hikey970.
Rename DRM driver kirin960 to kirin9xx to be compatible with hikey960 and hikey970.
