+++ drivers/gpu/drm/hisilicon/Makefile
@@ -3,4 +3,4 @@
 # Please keep this list sorted alphabetically
 
 obj-$(CONFIG_DRM_HISI_KIRIN) += kirin/
-obj-$(CONFIG_DRM_KIRIN_960) += kirin960/
+obj-$(CONFIG_DRM_KIRIN_960) += kirin9xx/
