+++ drivers/gpu/drm/hisilicon/kirin9xx/dw_dsi_reg.h
