+++ drivers/gpu/drm/hisilicon/kirin9xx/hdmi/adv7535.c
