+++ drivers/gpu/drm/hisilicon/kirin9xx/hdmi/adv7535_audio.c
