+++ drivers/gpu/drm/hisilicon/kirin9xx/kirin_drm_dpe_utils.h
